+++ ante/cosmos/eip712.go
@@ -1,6 +1,7 @@
 package cosmos
 
 import (
+	"context"
 	"fmt"
 	"strconv"
 
@@ -12,7 +13,10 @@ import (
 	"github.com/cosmos/evm/crypto/ethsecp256k1"
 	"github.com/cosmos/evm/ethereum/eip712"
 
+	txv1beta1 "cosmossdk.io/api/cosmos/tx/v1beta1"
 	errorsmod "cosmossdk.io/errors"
+	txsigning "cosmossdk.io/x/tx/signing"
+	"cosmossdk.io/x/tx/signing/aminojson"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -21,11 +25,10 @@ import (
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authante "github.com/cosmos/cosmos-sdk/x/auth/ante"
-	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 )
 
-var evmCodec codec.ProtoCodecMarshaler
+var evmCodec codec.Codec
 
 func init() {
 	registry := codectypes.NewInterfaceRegistry()
@@ -177,17 +180,35 @@ func VerifySignature(
 			return errorsmod.Wrap(errortypes.ErrNoSignatures, "tx doesn't contain any msgs to verify signature")
 		}
 
-		txBytes := legacytx.StdSignBytes(
-			signerData.ChainID,
-			signerData.AccountNumber,
-			signerData.Sequence,
-			tx.GetTimeoutHeight(),
-			legacytx.StdFee{
-				Amount: tx.GetFee(),
-				Gas:    tx.GetGas(),
+		anyMsgs, err := eip712.ToAnyMsgs(msgs)
+		if err != nil {
+			return err
+		}
+		feeAmount := eip712.ToFeeAmount(tx.GetFee())
+		txData := txsigning.TxData{
+			Body: &txv1beta1.TxBody{
+				Messages:      anyMsgs,
+				Memo:          tx.GetMemo(),
+				TimeoutHeight: tx.GetTimeoutHeight(),
 			},
-			msgs, tx.GetMemo(),
-		)
+			AuthInfo: &txv1beta1.AuthInfo{
+				Fee: &txv1beta1.Fee{
+					Amount:   feeAmount,
+					GasLimit: tx.GetGas(),
+				},
+			},
+		}
+		signModeHandler := aminojson.NewSignModeHandler(aminojson.SignModeHandlerOptions{})
+		signer := txsigning.SignerData{
+			ChainID:       signerData.ChainID,
+			AccountNumber: signerData.AccountNumber,
+			Sequence:      signerData.Sequence,
+			Address:       signerData.Address,
+		}
+		txBytes, err := signModeHandler.GetSignBytes(context.Background(), signer, txData)
+		if err != nil {
+			return errorsmod.Wrap(err, "failed to get sign bytes using aminojson")
+		}
 
 		signerChainID, err := strconv.ParseUint(signerData.ChainID, 10, 64)
 		if err != nil {