+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [\#811](https://github.com/cosmos/evm/pull/811) Use sdk's DefaultBondDenom for default evm denom in genesis.
 - [\#823](https://github.com/cosmos/evm/pull/823) Remove authz dependency from test suite and EvmApp interface
 - [\#829](https://github.com/cosmos/evm/pull/829) Seperate test app interface
+- [\#860](https://github.com/cosmos/evm/pull/860) Repalce legacytx.StdSignBytes with the aminojson sign mode handler in eip712.
 
 ### FEATURES